Create the Ship and Gameboard factories
The Ship factory creates a ship that can be used by the Gameboard
to place it anywhere on the board. The Gameboard keeps track of
hits and whether it hit a ship or not, calling the respective
ship's hit() function if so. It keeps track of all the ships and their
positions on the board and ensures that their placements are legal.